requirements.txt
absl-py==1.0.0
altgraph==0.17.2
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.5.2
aspose-words==22.8.0
astunparse==1.6.3
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
binance-connector-python==0.6.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.4
Django==4.1.1
django-autocomplete-light==3.9.4
django-bootstrap-datepicker-plus==5.0.0
django-bootstrap-v5==1.0.11
django-bootstrap4==22.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-rest-knox==4.2.0
django-rosetta==0.9.8
django-settings-export==1.2.1
django-simple-history==3.2.0
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.7.1
Flask==2.2.2
flatbuffers==2.0
fonttools==4.28.5
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
hyperlink==21.0.0
idna==3.3
imgkit==1.2.2
importlib-metadata==4.10.1
incremental==21.3.0
ipykernel==6.6.0
ipython==7.30.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
newsapi-python==0.2.6
num2words==0.5.6
numpy==1.22.0
oauthlib==3.1.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
p2pnetwork==1.1
packaging==21.3
pandas==1.3.5
parso==0.8.3
passlib==1.7.3
pdfkit==1.0.0
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.2
polib==1.1.0
prompt-toolkit==3.0.24
protobuf==3.19.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.11.1
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pyOpenSSL==21.0.0
pyparsing==3.0.9
PyPDF2==1.26.0
pypiwin32==223
pyserial==3.4
pyTelegramBotAPI==4.2.2
python-dateutil==2.7.3
python-stdnum==1.13
python-telegram-bot==13.10
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyusb==1.0.2
pywin32==304
pywin32-ctypes==0.2.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
suds==1.1.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.4.0
tzdata==2022.2
tzlocal==4.1
urllib3==1.26.5
virtualenv==20.15.0
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.13.3
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0



astroid
black
cfgv
click
flake8
gunicorn
identify
isort
lazy-object-proxy
mccabe
mypy
mypy-extensions
nodeenv
pathspec
pre-commit
psycopg2-binary
pyflakes
pylint
pylint-django
pylint-plugin-utils
PyYAML
regex
typed-ast
typing-extensions
